feat: restructure sign-in and layout components; remove old files and implement new layouts with Clerk integration
