+++ src/app/(with-navbar)/sign-in/[[...sign-in]]/page.tsx
@@ -3,7 +3,7 @@ import { SignIn } from '@clerk/nextjs';
 export default function Page() {
     return (
         <div className="flex flex-col items-center justify-center h-screen bg-[var(--background)] dark">
-            <SignIn />
+            <SignIn forceRedirectUrl={'/dashboard'} />
         </div>
     );
 }
