+++ src/app/(without-navbar)/dashboard/page.tsx
@@ -13,6 +13,7 @@ import {
     SidebarProvider,
     SidebarTrigger,
 } from '@/components/ui/sidebar';
+import { UserButton } from '@clerk/nextjs';
 
 export default function Page() {
     return (
@@ -38,6 +39,9 @@ export default function Page() {
                             </BreadcrumbItem>
                         </BreadcrumbList>
                     </Breadcrumb>
+                    <div className="ml-auto">
+                        <UserButton />
+                    </div>
                 </header>
                 <div className="flex flex-1 flex-col gap-4 p-4">
                     <div className="grid auto-rows-min gap-4 md:grid-cols-3">
