+++ src/app/(without-navbar)/layout.tsx
@@ -1,0 +1,47 @@
+import type { Metadata } from 'next';
+import { Geist, Geist_Mono } from 'next/font/google';
+import '../globals.css';
+import { ClerkProvider } from '@clerk/nextjs';
+import { dark } from '@clerk/themes';
+
+const geistSans = Geist({
+    variable: '--font-geist-sans',
+    subsets: ['latin'],
+});
+
+const geistMono = Geist_Mono({
+    variable: '--font-geist-mono',
+    subsets: ['latin'],
+});
+
+export const metadata: Metadata = {
+    title: 'Create Next App',
+    description: 'Generated by create next app',
+};
+
+export default function RootLayout({
+    children,
+}: Readonly<{
+    children: React.ReactNode;
+}>) {
+    return (
+        <ClerkProvider
+            appearance={{
+                baseTheme: dark,
+                variables: {
+                    colorPrimary: '#14b8a6',
+                    colorBackground: '#000000',
+                },
+            }}
+        >
+            <html lang="en" className="dark">
+                <body
+                    className={`${geistSans.variable} ${geistMono.variable} antialiased`}
+                >
+                    <div className="relative w-full flex items-center justify-center"></div>
+                    {children}
+                </body>
+            </html>
+        </ClerkProvider>
+    );
+}
